feat(editor): Use text editor
